rename features folder into game
