+++ src/game/mod.rs
@@ -1,12 +1,9 @@
 pub(crate) mod cards;
 pub(crate) mod deck;
-//pub mod screen;
+pub(crate) mod screen;
 
 use bevy::prelude::*;
 
 pub(super) fn plugin(app: &mut App) {
-    app.add_plugins((
-        deck::plugin,
-        //dev_tools::plugin,
-    ));
+    app.add_plugins((deck::plugin, screen::plugin));
 }
